cohttp_lwt_unix: make benchmark more realistic
This change is based on https://github.com/inhabitedtype/httpaf/issues/200
in which @vouillon reasonably points out that it is unrealistic to assume that the request handler responds immediately.
The benchmark is modified to yield inside the request handler in order to simulate the effect of an hypotetical db request.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>